requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.28.2
jinja2==3.1.2
reportlab==3.6.12
pandas==2.0.0
openpyxl==3.1.2
slowapi==0.1.9
werkzeug==2.3.7
httpx==0.24.1
